requirements.txt
aiohttp==3.8.4
PyYAML==6.0
websockets==11.0